react-app/server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyMuPDF==1.23.8
pandas==2.1.1
openai==1.3.5
google-generativeai==0.3.2
matplotlib==3.7.2
numpy==1.24.3
requests==2.31.0
Werkzeug==2.3.7
